mostly making the code readable. change forced flush to only be activ… (#5)
* mostly making the code readable. change forced flush to only be activated if compiled for windows, where it actually matter.

* Fixed DNDEBUG define typo. More Readability. Fixed more linter issues: mostly casting-related, but also unused vars

* Accidentally an entire file

* more cleaning, and notes about bugs that can be fixed

+++ src/p25_lcw.c
@@ -1,8 +1,6 @@
 #include "dsd.h"
 
-void
-processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, char *lcinfo)
-{
+void processP25lcw(dsd_opts *opts, dsd_state *state, char *lcformat, char *mfid, char *lcinfo) {
 
     char tgid[17], tmpstr[255];
     long talkgroup, source;
@@ -10,24 +8,19 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
 
     tgid[16] = 0;
 
-    if (opts->p25lc == 1)
-    {
+    if (opts->p25lc == 1) {
         fprintf(stderr, "lcformat: %s mfid: %s lcinfo: %s ", lcformat, mfid, lcinfo);
-        if (opts->p25tg == 0)
-        {
+        if (opts->p25tg == 0) {
             fprintf(stderr, "\n");
         }
     }
 
-    if (strcmp (lcformat, "00000100") == 0)
-    {
+    if (strcmp(lcformat, "00000100") == 0) {
 
         // first tg is the active channel
         j = 0;
-        for (i = 40; i < 52; i++)
-        {
-            if (state->tgcount < 24)
-            {
+        for (i = 40; i < 52; i++) {
+            if (state->tgcount < 24) {
                 state->tg[state->tgcount][j] = lcinfo[i];
             }
             tmpstr[j] = lcinfo[i];
@@ -38,25 +31,21 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
         tmpstr[14] = 48;
         tmpstr[15] = 48;
         tmpstr[16] = 0;
-        talkgroup = strtol (tmpstr, NULL, 2);
-        state->lasttg = talkgroup;
-        if (state->tgcount < 24)
-        {
+        talkgroup = strtol(tmpstr, NULL, 2);
+        state->lasttg = (int) talkgroup;
+        if (state->tgcount < 24) {
             state->tgcount = state->tgcount + 1;
         }
-        if (opts->p25tg == 1)
-        {
+        if (opts->p25tg == 1) {
             fprintf(stderr, "tg: %li ", talkgroup);
         }
 
-        if (opts->p25tg == 1)
-        {
+        if (opts->p25tg == 1) {
             fprintf(stderr, "tg: %li ", talkgroup);
 
             // the remaining 3 appear to be other active tg's on the system
             j = 0;
-            for (i = 28; i < 40; i++)
-            {
+            for (i = 28; i < 40; i++) {
                 tmpstr[j] = lcinfo[i];
                 j++;
             }
@@ -65,11 +54,10 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
             tmpstr[14] = 48;
             tmpstr[15] = 48;
             tmpstr[16] = 0;
-            talkgroup = strtol (tmpstr, NULL, 2);
+            talkgroup = strtol(tmpstr, NULL, 2);
             fprintf(stderr, "%li ", talkgroup);
             j = 0;
-            for (i = 16; i < 28; i++)
-            {
+            for (i = 16; i < 28; i++) {
                 tmpstr[j] = lcinfo[i];
                 j++;
             }
@@ -78,11 +66,10 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
             tmpstr[14] = 48;
             tmpstr[15] = 48;
             tmpstr[16] = 0;
-            talkgroup = strtol (tmpstr, NULL, 2);
+            talkgroup = strtol(tmpstr, NULL, 2);
             fprintf(stderr, "%li ", talkgroup);
             j = 0;
-            for (i = 4; i < 16; i++)
-            {
+            for (i = 4; i < 16; i++) {
                 tmpstr[j] = lcinfo[i];
                 j++;
             }
@@ -91,20 +78,14 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
             tmpstr[14] = 48;
             tmpstr[15] = 48;
             tmpstr[16] = 0;
-            talkgroup = strtol (tmpstr, NULL, 2);
+            talkgroup = strtol(tmpstr, NULL, 2);
             fprintf(stderr, "%li\n", talkgroup);
         }
-    }
-
-    else if (strcmp (lcformat, "00000000") == 0)
-    {
+    } else if (strcmp(lcformat, "00000000") == 0) {
         j = 0;
-        if (strcmp (mfid, "10010000") == 0)
-        {
-            for (i = 20; i < 32; i++)
-            {
-                if (state->tgcount < 24)
-                {
+        if (strcmp(mfid, "10010000") == 0) {
+            for (i = 20; i < 32; i++) {
+                if (state->tgcount < 24) {
                     state->tg[state->tgcount][j] = lcinfo[i];
                 }
                 tmpstr[j] = lcinfo[i];
@@ -114,13 +95,9 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
             tmpstr[13] = 48;
             tmpstr[14] = 48;
             tmpstr[15] = 48;
-        }
-        else
-        {
-            for (i = 16; i < 32; i++)
-            {
-                if (state->tgcount < 24)
-                {
+        } else {
+            for (i = 16; i < 32; i++) {
+                if (state->tgcount < 24) {
                     state->tg[state->tgcount][j] = lcinfo[i];
                 }
                 tmpstr[j] = lcinfo[i];
@@ -128,33 +105,27 @@ processP25lcw (dsd_opts * opts, dsd_state * state, char *lcformat, char *mfid, c
             }
         }
         tmpstr[16] = 0;
-        talkgroup = strtol (tmpstr, NULL, 2);
-        state->lasttg = talkgroup;
-        if (state->tgcount < 24)
-        {
+        talkgroup = strtol(tmpstr, NULL, 2);
+        state->lasttg = (int) talkgroup;
+        if (state->tgcount < 24) {
             state->tgcount = state->tgcount + 1;
         }
-        if (opts->p25tg == 1)
-        {
+        if (opts->p25tg == 1) {
             fprintf(stderr, "tg: %li ", talkgroup);
         }
 
         j = 0;
-        for (i = 32; i < 56; i++)
-        {
+        for (i = 32; i < 56; i++) {
             tmpstr[j] = lcinfo[i];
             j++;
         }
         tmpstr[24] = 0;
-        source = strtol (tmpstr, NULL, 2);
-        state->lastsrc = source;
-        if (opts->p25tg == 1)
-        {
+        source = strtol(tmpstr, NULL, 2);
+        state->lastsrc = (int) source;
+        if (opts->p25tg == 1) {
             fprintf(stderr, "src: %li emr: %c\n", source, lcinfo[0]);
         }
-    }
-    else if ((opts->p25tg == 1) && (opts->p25lc == 1))
-    {
+    } else if ((opts->p25tg == 1) && (opts->p25lc == 1)) {
         fprintf(stderr, "\n");
     }
 }
